agents/schoopet/requirements.txt
google-adk[agent-identity]==1.33.0
google-genai==1.73.1
google-cloud-aiplatform==1.151.0
google-cloud-firestore==2.27.0
google-cloud-secret-manager==2.28.0
google-cloud-tasks==2.22.0
httpx==0.28.1
pydantic==2.13.4
cloudpickle==3.1.2
pytest==9.0.3
pytest-asyncio==1.3.0
pypdf==6.11.0
python-dotenv==1.2.2
pyOpenSSL==26.2.0
google-api-python-client==2.196.0
opentelemetry-instrumentation-google-genai>=0.4b0
opentelemetry-instrumentation-sqlite3
opentelemetry-exporter-gcp-logging
opentelemetry-exporter-gcp-monitoring
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-vertexai>=2.0b0
